Moved the configurations for Monitoring, Serialization, and Routing plugins to individual files in the 'plugin' folder for better organization.
