+++ src/main/kotlin/io/sebi/plugins/Monitoring.kt
@@ -1,0 +1,13 @@
+package io.sebi.plugins
+
+import io.ktor.server.application.*
+import io.ktor.server.plugins.callloging.*
+import io.ktor.server.request.*
+import org.slf4j.event.Level
+
+fun Application.configureMonitoring() {
+    install(CallLogging) {
+        level = Level.INFO
+        filter { call -> call.request.path().startsWith("/") }
+    }
+}
