+++ src/main/kotlin/io/sebi/plugins/Serialization.kt
@@ -1,0 +1,11 @@
+package io.sebi.plugins
+
+import io.ktor.serialization.kotlinx.json.*
+import io.ktor.server.application.*
+import io.ktor.server.plugins.contentnegotiation.*
+
+fun Application.configureSerialization() {
+    install(ContentNegotiation) {
+        json()
+    }
+}
